+++ frontend/components/ui/mention/mention-view/MentionView.tsx
@@ -5,20 +5,13 @@ import { MENTION_MATCH_PATTERN } from '@/constants/mention-match-pattern.const';
 import { MentionsResponse } from '@/shared/interfaces/mention.interface';
 import { AnimatePresence, motion } from 'framer-motion';
 import MemberHoverModal from '../../modal/member-hover-modal/MemberHoverModal';
+import { useHover } from '@/hooks/useHover';
 
 const MentionView: FC<{ contents: string; mentions: MentionsResponse[] }> = ({
 	contents,
 	mentions,
 }) => {
-	const [isHovering, setIsHovering] = useState<number>();
-
-	const handleMouseOver = (index: number) => {
-		setIsHovering(index);
-	};
-
-	const handleMouseOut = (index: number) => {
-		setIsHovering(undefined);
-	};
+	const { handleMouseOver, handleMouseOut, isHovering } = useHover();
 
 	const renderMentions = (text: string) => {
 		const parts = text.split(MENTION_MATCH_PATTERN);
